feat: add ens and fc data to optimistic comment
